Better GRPC support
With these additions it’s the first way to make better use of Displaying WebSockets for Streaming Requests in general for gRPC Requests.

I did some breaking changes to determine the usage of unary calls and streaming calls to be better seperated.

In the Readme it’s even provided how to write a proper interecptor for working with that.